fix(homework-5): PR comments

+++ sensors/Sensor.h
@@ -21,7 +21,7 @@ namespace SmartHome {
 
     protected:
         std::string name;
-        int *value;
+        double *value;
         MeasureUnit unit;
     };
 }
